Use a proper error code for "attribute X has wrong type" errors.
GPDB appends source file name and line number to internal errors, so
all user-facing errors must have a proper error code. That is the upstream
policy as well, but upstream isn't always as careful about it. (This
case was fixed in upstream in commit 09cecdf285, this is a partial
cherry-pick of that.)