Reuse active snapshot instead of catalog snapshot when possible
Catalog snapshot is a replacement for SnapshotNow and should be used
only when scanning catalog tables.  gp_distribution_policy extension
is used to scan a user table, so use active snapshot instead.  In the
AO format version UDF, active snapshot is suffice as it's invoked by a
test where concurrent updates to the tuple in pg_appendonly are not
expected.