Remove specific error message for v0 API functions.
We're seeing that error in the mapreduce tests in the pipeline, but the
old error message didn't include the function name, so I'm not 100% sure
which function was triggering it.

The support for V0 API was also removed in upstream, in PostgreSQL v10
(commit 5ded4bd214). So one can think of this as a backport of that.