Use GetCatalogSnapshot() instead of GetLatestSnapshot() in some places.
Many of the scans on AO table metadata need a more recent snapshot than
what GetLatestSnapshot() returns. Especially with repeatable read or
serializable isolation level, the latest snapshot can be quite old.
Otherwise, you might overwrite changes to the EOF position in an AO
segment, for example. This explains many of the AO related errors in the
regression tests.