Revert some changes to 'horology', 'misc' tests, made while fixing conflicts.
We stil need to drop the GPDB-specific horology test tables. Or if not, we
need to include them in the 'misc' test's output of all user tables. Seems
better to drop them, they're not interesting from a dump & restore testing
point of view either, because they're identical to tables created by other
tests.

I'm not sure why we use bespoken tables in the horology test in GPDB, rather
than rely on the tables created by the datatype-specific tests, like in
upstream. But I'm not going to try changing that as part of the merge.