Fix pg_resqueue_status view.
In the merge, the LWLock functions changed slightly, and the code in
BuildQueueStatusContext() that acquires all the partition-locks was
changed. As part of that, we got confused on what the 'i' variable was
used for. It was previously used for the current record number, but after
the merge, it was left at NUM_LOCK_PARTITIONS (i.e. 16), and the counting
of record numbers started from that. Fix the confusion.