Re-enable SRF-in-targetlist support for Agg nodes.
We've removed SRF-in-targetlist support for most node types in GPDB.
As a micro-optimization, I believe. Instead, we insert Result nodes to
evaluate the set-returning-functions. That's a good tradeoff, because
SRFs-in-targetlist are rare.

To fix, put back the handling of SRF-in-targetlist to Agg node, as it is
in the upstream.

This fixes a couple of cases already exercised by the regression tests,
where we had just memorized the error.

This came up in the 9.4 merge again, which added another such query,
in the 'json' test. Seems like a good time to fix this properly.

PS. PostgreSQL v10 changed things around this, so that there is a new
ProjectSet node type to evaluate SRFs. So that's similar to what we're
doing in GPDB, but in a less quick & dirty way. I don't feel like trying
to backport that, though, it was a pretty large patch, and there very
several follow-on fixes. So let's try to limp on with the current system,
until v10 merge.

NOTE: Cherry-picked from https://github.com/greenplum-db/gpdb/pull/5917.
Make sure to update iteration branch, if there are changes to that PR.