Don't print disabled internal FK triggers in psql output.
Foreign key triggers are always disabled in GPDB, so it seems pointless.
It might be nice to somehow highlight in the psql output the fact that
foreign keys are not enforced in GPDB, perhaps by adding "(disabled)" to
the list of foreign keys. But printing the names of the internal triggers
is probably not the way we want to do that.