Function Scans can be rescanned, don't eager free too eagerly.
Function Scan materializes the result of each in a TupleStore, and can
be rescanned. Mark it as rescannable in the planner.

If the Function Scan contains multiple functions, then don't free it
eagerly. That way, the result of some functions can be reused, if it's
rescanned. This fixes the failure in the 'rangefuncs' test.

NOTE: This could be done as a separate PR.