Silence 'partition_locking' failure.
These are all cases where we now hold locks in segments longer than we used
to. We didn't mean to change these in the merge, but I don't see any good
reason to play fast and loose with locking in these cases. So accept the
stronger locks.

NOTE: This should probably be investigated further. See also
https://github.com/greenplum-db/gpdb/issues/5919.