Work around the duplicate-NOTICE elimination of gpdiff.
gpdiff removes duplicate NOTICEs coming from segments. That was causing
a regresson failure in upstream tests that throw NOTICEs. In the upstream
expected output, the NOTICEs have no segment information, but in the
actual GPDB output, they do. gpdiff therefore removed the duplicates from
the actual output, but not from the expected output, which made the
comparison fail.

Adjust expected output with GPDB's segment information, to work around
gpdiff's cleverness. Unfortunately this loses some coverage: if the
tests incorrectly printed a different number of NOTICEs, for some reason,
we would mask out the difference. We could perhaps add some extra flags
for gpdiff or something, to give it the behavior we want, but I think
this will do for now.