Try harder to find security_barrier views.
RangeTblEntry->security_barrier isn't set until the expand_security_quals()
call in the planner. That never happens in ORCA, so we can't rely on it.
Furthermore, I think the old check was only performed on views that
appeared in the FROM clause, not for UPDATE/DELETE targets. Move the check,
so that it's performed on all relations in the range table. That might be
overly conservative, if e.g. a security barrier view's relations are
eliminated by constraint exclusion, but better to err on the side of
caution.