Adjust expected output for missing "Vmem reserved" lines.
I did some debugging on why we don't get the "Vmem reserved" on slice 1
in these queries anymore. I'm still not 100% sure where exactly, but I'm
convinced it's harmless. I think the amount of memory used in that slice
is just on the edge of whether it needs to reserve a "chunk" or not, and
some unrelated small change in the merge brought it just below the
threshold. Testing on my laptop, it seems that if the query is the very
first one in the session, then you do get the "Vmem reserved", and
debugging shows that it's because setting up the interconnect requires
some memory on the first call. On subsequent calls, the interconnect
receive buffer has already been allocated, and the total allocations stay
just below the threshold. On 'master', the allocations that trigger the
chunk to be allocated are from building a catcache entry, and from
allocating the buffer for uncompressing the serialized plan. That all
seems very random, so I think we can just accept the new expected
output.

It would be good to make this test more robust, perhaps by using a
different query. But this will do for now.