Take AccessExclusiveLock in ALTER TABLE operations that add/remove partitions
Upstream relaxed the locking of various ALTER TABLE commands, but I don't
think we're ready to do that for GPDB's additional ALTER TABLE variants.
Maybe we could for some, but that would need further investigation. pg_upgrade
was failing, because ALTER TABLE command to add subtemplates didn't hold
an AccessExclusiveLock while it tried to create a toast table.