Use bootstrap superuser as the authenticated user in GDD process.
That's what we were effectively choosing for session user before already
but findSuperuser() was an overly complicated way to do it. In addition
to setting the session userid, also set authenticated userid. After the
previous commit, we're now relying on it to be set when allocating a gang.

NOTE: This is a nice cleanup, should open a separate PR.