Remove the stale FOREIGN_KEY relationship of pg_class.
This was causing gpcheckcat to not build the correct relationship using the json
file. Which silently failed internally and hence other tables in gpcheckcat
after pg_class didn't get the foreign relationship as well. That caused the
behave gpcheckcat complain for tags @foreignkey_full_segment.

The reltoastidxid was renamed to reltoastrelid in 9.4.