Update flaky FTS isolation2 tests
Updated tests to call `gp_fts_probe_scan()` twice to avoid overlap with
auto-trigger of FTS scans by FTS process. As if that happens, due to race
condition, it will not trigger the fault and fail the test.

The issue exists on master too. We will create a seperate PR on master for it.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>