Reboot our attempts at making 'eplain_format' output stable.
I don't think we need to test all four formats with a complicated
EXPLAIN ANALYZE query. The XML, JSON and YAML outputs should be the same,
except for the formatting, so if we test the text format, and one of
the tree formats, that should be enough. I kept the YAML case, and
replaced the XML and JSON tests with a much simpler query, with ANALYZE
off. Those very basic tests should be enough to verify that the formatting
basically works, and the more complicated YAML test covers that the tree
formats include all the information.

I also removed the gpdiff matchsub rules for the slice numbers. I don't
think we should obfuscate those. These queries should always have the
same number of slices, and if that ever changes, we can update the
expected output.

I don't think this is enough to solve the flakiness with the "Vmem
reserved" stuff yet. I expect we'll still see failures because of that in
the pipeline. But now that we have fewer copies of the same thing, we have
a better chance of dealing with that with having enough alternative
expected output files to cover all the cases. I'm going to add those,
after the pipeline has run on this once, and we know what alternative
outputs we need.