A tentative fix for panic caused by an implicit lateral join.
The related sql is as below in the rangefuncs regression test.

select x from int8_tbl, extractq2_2(int8_tbl) f(x);

Current plan is:

explain (verbose, costs off)
select x from int8_tbl, extractq2_2(int8_tbl) f(x);
                   QUERY PLAN
------------------------------------------------
 Nested Loop
   Output: ((int8_tbl.*).q2)
   ->  Gather Motion 3:1  (slice1; segments: 3)
         Output: int8_tbl.*
         ->  Seq Scan on public.int8_tbl
               Output: int8_tbl.*
   ->  Materialize
         Output: ((int8_tbl.*).q2)
         ->  Result
               Output: (int8_tbl.*).q2
 Optimizer: legacy query optimizer
(11 rows)

Note pg upstream does not have the unncessary materlized node
because there is originally a subquery node above the result
node but the subquery scan node is eliminated at the set reference
stage (very late) and GPDB will add a materialize node above
the subquery node in the nestloop join path if needed
(!rescannable). This could be improved for our case of course
but not in this patch.

The panic is because that the result node does not have outer
plan and it gets the tuple from the parameter via FieldSelect
- that of course conflicts with inner prefetch (in nestloop
join) which was enabled since the inner path node of the
nestloop join node is a materialize node.

This patch is not a perfect fix, but seems to be a good workaround.
We might want further work when we finally support lateral gracefully.