Partially fix join.
For those cases which are intended for merge join test, specify
proper GUCs to enforce merge join.

Also for the case that tests extraction of restriction OR clauses
from join OR clause, we do not limit it to bitmap index scan which
is known to be best path only when some GUCs are used because of
the difference of some settings between gp and gp.