pg_upgrade: explicitly perform vacuum freeze for template0.
After initdb, relfrozenxid for pg_statistic table is set to 3. But the
xmin for the rows still have transaction id > 3. When pg_upgrade
directly updates the relfrozenxid for pg_statistic to something like 800
and performs vacuum freeze, encounters the error like "ERROR: found xmin
570 from before relfrozenxid 739".

Instead of carrying the assumption for initdb to do the job better for
pg_upgrade to also vacuum freeze template0 and avoid hitting this error.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>