Resove conflicts in src/bin/pg_rewind (except parsexlog.c)
The fix is stright-forword, 95% of work is just keep the HEAD change.

parsexlog.c is much different with upstream because 686145bd532 (GP) change
it and use 9.4 xlog parsing logic.

The next step is revert 686145bd532 and make sure the changes in 0a08f3195776d (GP)
and 97a76214c5897 (GP) and 161920e869a3 (GP) is also applied.