Resolve conflicts in src/backend/replication/logical/decode.c
There were several conflicts due to backports to 9.4 STABLE on top of
code feature changes between 9.4 and 9.5. As an example backport
difference, (postgres) commit 3b94b3a496f in 9.4 STABLE is different
than the corresponding 9.5 commit 5990a034adf because of revamp of WAL
record format as exampled in (postgres) commit 2c03216d831. Additionally
there was a commit only in 9.4 STABLE which introduced a code difference
in DecodeUpdate in how record length was calculated which now needs to
be removed in this merge.