Resolve conflicts in src/bin/pg_rewind/parsexlog.c
The main conflict is located in extractPageInfo(), it was modified to
work with 9.4 wal record format, now the wal format is merged to match
9.5, the logic should also match upstream 9.5.

In addition, RM_APPEND_ONLY_ID is specially treated because it needs a
different call to process_aofile_change(). For RM_BITMAP_ID which is also
a GPDB-specified type, we don't treat it as a special in extractPageInfo(),
we hope the buffers are registered properly when xlog is inserted.