Resolve last conflicts in access/heap/heapam.c
Greenplum has LockTupleWaitType which seems to be the same as the
newly introduced LockWaitPolicy (df630b0dd5ea2de5). They look very
similar so take upstream Postgres's LockWaitPolicy. This will need
more investigation in the near future so we added a couple FIXME
comments that can be addressed later when we have the database up and
running.

Co-authored-by: David Kimura <dkimura@pivotal.io>