Resolve conflicts in src/backend/utils/sort
The main conflicts comes from upstream commit cdd5178c69 which extend the
MinimalTuple concept to tuplesort.c to reducing the per-tuple space. The
problem is GPDB and upstream store MinimalTuple in a different way, GPDB
use MemTuple as underlying storage, if we treat sort tuple as MinimalTuple
like upstream, it will be problematic, so this commit keep the GPDB way.