Resolve conflicts in vacuum and refactor it
This required a lot more work to resolve. A simple change in upstream
Postgres to no longer pass the VacuumStmt through vacuum was actually
very disruptive for Greenplum vacuum. Greenplum vacuum has a lot of
extra stuff in VacuumStmt (a flag to skip twophase commit and other
variables that control vacuum for AO tables). The decision we made was
to follow upstream Postgres's lead and also stop passing the
VacuumStmt. To do this, we had to create a new struct
AOVacuumPhaseConfig and pass our custom VacuumStmt variables
around. We also now construct a VacuumStmt before every vacuum
dispatch which is fine since we did that before for toast and AO
auxiliary tables anyways.

Postgres Reference:
0d831389749a3baaced7b984205b9894a82444b9