Adjust the MAX_MAPPINGS for relmaps
In Postgres this is 62, but Greenplum has exceeded this number due to
all of the extra Greenplum specific shared relations. Is it safe to bump
this up to 126 to occupy exactly 1 kilobyte? According to the internet,
modern disks nowadays have 4 kilobytes fixed sector size.

Bump it up regardless to get initdb working. We can review the fixme
later to see how safe this is.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>