Fix ALTER TABLE ALTER COLUMN [SET|DROP] DEFAULT
The code that is being removed is a small part of a patch back in 2012
to fix attrdef oid inconsistencies. This was redone in during the 5X
dev cycle in
https://github.com/greenplum-db/gpdb/commit/5cdad61ee27cf17b. Revert
the last pieces back to what they currently look like in upstream
Postgres 9.5.

This was caught while fixing the alter_table regress test.