Fix DROP POLICY crash
The policy relation was being closed before cache invalidation which
is super incorrect. Upstream postgres fixes this so backport the
majority of that fix commit. We don't cherrypick so that the commit
with the extra changes can apply better in future merges (I think
that's how git works).

Upstream Postgres commit:
https://github.com/postgres/postgres/commit/c244a511ba7

Co-authored-by: David Kimura <dkimura@pivotal.io>