Dispatch CREATE POLICY and CREATE TRANSFORM
The Oids to dispatch were not being created because they were not a
part of the oid_dispatch catalog list that we maintain. The CREATE
functions were not explicitly dispatching either. We also needed to
add corresponding node types to handle the dispatch on the primaries.

This commit also fixes object_address regress test which is where this
issue was found.