Extend atmsort.pm to sort NOTICE/WARNING messages of queries to aid gpdiff
When multiple QEs report NOTICE/WARNING messages to QD, the order
of those messages in QD are undetermined, this is painful to write
answer files for such test cases. To resolve this problem, we extend
the atmsort.pm script to sort NOTICE/WARNING messages of queries in
results files and answer files before do the differ.