Fix vacuum_full_heap regress test
VACUUM FULL does not update the relpages and reltuples of a relation's
toast table. This is because the relation and its toast table are
rebuilt via cluster_rel.

VACUUM FULL on GPDB6 will update the relpages and reltuples of a
relation's toast table local to the segment because the QD dispatches an
unnecessary VACUUM to the newly created toast table. Since the stats are
not being returned to the QD anyways, it seems unnecessary to record the
relpages and reltuples for the toast table. This was improperly managed
in the 9.0 merge.

We are not losing coverage by removing the checks on the relpages and
reltuples because we are still running pg_relation_size() on the toast
table during each phase (e.g. before/after table deletions and vacuum).

Co-authored-by: Jimmy Yih <jyih@pivotal.io>