Fix oid_wraparound regress test
When the test was created, the last scenario expected the nextOid for
QD to be 16388. During the 9.4 merge, postgres commit 5b3ed6b7880b was
introduced which prevented getting a new Oid that collided with
recently-dead entries so the expected nextOid was changed to 16389. In
this 9.5 merge, the colliding tuple is pruned out during the index
scan on pg_type so there is no more collision. Changed the test to
accept either 16388 or 16389.