Fix CTAS tuple count output
When running CTAS, it would output an incorrect tuple count of how
many tuple were selected to insert into the table. This was because of
an incorrect merge conflict resolution where the
queryDesc->es_processed was being recorded before the executor got the
results back from dispatch which happens in ExecutorEnd().