Fix issue that expanded objects not be handled by datumCopyWithMemManager
GPDB has its own version of datumCopy (datumCopyWithMemManager) which is
used to acceleration aggregation. In upstream, datumCopy palloc new memory
to copy the new transValue to it and pfree the old transValue memory. In
GPDB, if memory of old transValue can hold the new transValue, the new
transValue will be copied to the memory of old transValue. datumCopy can
support expanded objects now, datumCopyWithMemManager should do the same.