Force ORCA to use target list of motion instead of outer node
GPDB commit ff8b6c638 made the projection of motions the target list of the outer
node. However, in ORCA the projection of the outer node does not include
the alias. We may want to fix this in ORCA, but for now we force ORCA to
use the target list of the motion and add a FIXME.

Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: David Kimura <dkimura@pivotal.io>