Fix crash in file_fdw test
Relax assert for valid item pointer. Includes FIXME to investigate
further as I am not confident this should be relaxed. It triggered when
running select query using row-level lock modes on a table that includes
a foreign table in the inheritance hierarchy. An example of this is seen
inside `file_fdw` test.