Remove GPDB_95_MERGE_FIXME in copy.c
In DoCopy(), if check_enable_rls() is true, the relation might be
closed temporarily, so in BeginCopyTo and BeginCopyFrom, they call
BeginCopy() which recheck whether the relation referenced by COPY
statement has changed or not. In GPDB, BeginCopyToOnSegment() and
BeginCopyToForExternalTable() are called after BeginCopyTo() and
BeginCopyFrom(), so it is unnecessary to do it again, passing a
InvalidOid should be ok for those two functions, remove the FIXME
added by e1ae432539d40ed.