Disable UPSERT clause DO UPDATE SET
In GPDB we disabled "DO UPDATE SET" clause because as-is it can create
data violations.  For example, suppose we have a table with PRIMARY KEY
column that is not the distribution key. UPDATE could SET duplicate
PRIMARY KEY value when an already existing tuple with PRIMARY KEY value
is stored on another segment.

In the future we can relax this restriction if we check that the UPDATE
is not on a PRIMARY KEY. An even more ambitious solution might involve
coordination betwen QEs when PRIMARY KEY is updated so as to ensure no
data violations.