Remove FIXME and update the code
We can safely set sameslice_relids, that indicate which base relations
will be executed in the same slice, on samplescan paths. It is an unlikely
optimization due to the nature of tablescans, however as far as paths are
concerned the same optimization that can apply to seqscans can apply to
samplescans.