Remove FIXME from tablesample tests and improve coverage
Initially the tablesamples tests were based on a skewed data distribution
on one segment in order for the tests to be deterministic and the test files
to not diverge too much from upstream. However, alread enough diversion has
been added in order to not support this notion.

In order to test tablesample in data present in all segments and still maintain
the test deterministic, the amount of inserted rows has been tripled and the
distribution key has a repeatable value that will sent the data to the
corresponding segment.