Fix reduce errors in gram.y after conflict resolution.
In 9.6 operator_def_elem was introduced which is based on def_arg and NONE.

The Initial Greenplum code dump contains NONE in the def_arg which is a
deviation from upstream. In the supplied comment the decision was taken in order to
avoid adding keyword. In any case, this causes a reduce/reduce error for the
newly added rule.

Rewrite the rule for now until a proper fix is introduced.

Adds a FIXME.