Resolve conflicts around planner.
The upper planner pathification wrecked havoc in planner.c. A lot of
GPDB specific code needs to be changed from working with Plans, to
working with Paths. But for now, I just moved existing code roughly to
where it needs to live in the future, without actually making it work.