Fix conflicts in src/backend/replication/logical.
I resolved these a bit lazily by just replacing with the REL9_6_STABLE
version. We got some of the later fixes already from an earlier point
release. We got some extra fixes as part of this, but that seems pretty
harmless, there was nothing big and I was too lazy to tear them apart,
just to get them again as part of a later point release.