Replace src/timezone/ with the version from tip of upstream REL9_6_STABLE.
Commit f4b91a50e9, to be precise.

We had some intermediate version of the timezone code and data, that didn't
correspond to any 9.6.X minor release that I could identify. It was more
recent than the point we're now merging with, though. Rather than try
fixing the conflicts manually, pick the upstream version.

I considered taking the upstream version that we are merging with - that
would avoid merge conflicts going forward - but that would revert us to
an older version than we already had, which seems strange. Picking this
newer version will cause some merge conflicts going forward, but since
there are no GDPB specific changes in the directory (except for the extra
gptime.c file), we can easily just replace the directory with the latest
version whenever we need to.