Fix conflicts in src/backend/nodes.
There was a lot of duplication betwee readfast.c and readfuncs.c now,
because upstream added read functions for all the Plan nodes. I think
we need to do more here; we need to move the read functions for all the
GPDB specific Plan type to readfuncs.c, too. Otherwise, they won't work
with parallel query. But for now, at least it compiles.