Schema-qualify stuff that needs to go to pg_catalog schema.
I'm not sure why this is now needed, or even 100% sure that it is, but
seems like a good idea in any case. A view in gp_toolkit were complaining
that pg_catalog.gp_pgdatabase doesn't exist, and this fixed it. (But
there are other issues in gp_toolkit, so keep it disabled for now)