Make sync rep work for QD
In GPDB, the FTS process is responsible for populating the value of
the synchronous_standby_names guc in the postgresql.auto.conf and then
signalling the postmaster to reload this configuration file to apply the
changes. FTS only works on the segments and thus does not perform the
above for QD.

As of commits 989be0 and 4c804f parsing of the postgresql.auto.conf file
is done with a separate gram.y file. At the end of parsing, the
SyncRepConfig struct is populated. This struct is used to determine the
list of sync standby names, which are now used to get the oldest sync
lsn values.

Since health-checking is not currently done for the QD, a value of '*'
does not make sense in QD's postgresql.auto.conf. A value of '*' demands
that the primary waits for at least one of the standbys to replicate. If
there are no standbys alive, the primary would wait forever. This may
change with the implementation of master standby auto-failover.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivtoal.io>