Accept plan changes for one- vs two-phase aggs.
The planner estimates that these queries return only a single row,
and in that case a one-phase aggregate is indeed marginally cheaper.
I checked that if there's more data in the table, the planner does choose
a two-phase agg like before.

Perhaps we should somehow nudge the planner to choose the two-phase agg
more aggressively. using a one-phase agg is a pretty risky bet, it will
lose out to the two-phase agg, if the estimate is wrong and there are in
fact many input rows. But it is logical as it is.