Add a test case for the two-stage aggs.
We have plenty of tests for these things already, but it seems handy to
cover all the different one-phase, two-phase, DQA, etc. grouping plan
types in one concise test. Also, I'm not sure we have EXPLAIN tests for
all of these, so I'm not sure if we would notice if we lost the ability
to create some of the plan types.

NOTE: We could put this to 'master' before the merge commit, all of these
queries should work on master already.