Serialize/deserialize Constraint->initially_valid
Even though initially_valid was added some time ago, until the NOT VALID
syntax was added in the merge, it seems like its value didn't matter.
Now, we use the value of initially_valid (initialized to true in the
parser) to set the value of skip_validation which will determine if we
skip validation for a particular tuple. This is happening correctly,
but, we forgot to add it to the deserialization/serialization for the
Constraint node, so the segments were not getting the memo.

Also, update remove conflict markers from the alter_table test files.

alter_table.sql does not pass now, however, this fixes one
failure--before this commit
ALTER TABLE tmp3 VALIDATE CONSTRAINT identity;
incorrectly printed out
NOTICE:  boo: 20
NOTICE:  boo: 18
NOTICE:  boo: 16
now it prints out only
NOTICE:  boo: 20
NOTICE:  boo: 16