Fix two-stage Limit.
The preliminary Limit was not added in some queries where it was before,
per the regression tests (e.g. 'inherit' test).

The previous plan in the 'subselect' actually looked better, the new one
with two Limit nodes on top of each other is silly. But I believe that
will get fixed by https://github.com/greenplum-db/gpdb/pull/8684, so just
accept the silly plan for now.