Fix "volatile EquivalenceClass has no sortref" errors with DISTINCT.
I'm not entirely sure why these changes are needed now,a nd not before.
It does make sense that if there is a volatile expression in the target
list, we must use the 'sortgroupref' to refer to it. Just like with
ORDER BY and GROUP BY expressions.

I'm feeling that this area would be ripe for some refactoring; it's
a bit awkward to build multiple exprs/opfamilies/sortref lists like this.
Maybe pass around lists of TargetEntrys instead, or something. But this
will do for now.