Accept plan change.
There's an extra Result node in the new plan. With EXPLAIN VERBOSE, you can
see that it project away some unnecessary columns from the targetlist,
before the Motion. So that makes sense, although I'm not sure why those
unnecessary columns are in the targetlist in the first place. They were
not there before the 9.6 merge. But this is such a rare corner case that
it doesn't seem worth investigating.