Reinstate catalog/gp_toolkit schema
Most of the failures were a consequence of empty new lines within DELIMITERS in
plgpsql to be parsed as query delimiters, triggering parse errors. Usually this
should have been avoided by the '-j' flag in command (postgres). We do pass the
flag and from a quick enough look, there does not seem to be something unusual
there. Regardless it is a small enough issue which imho does not affect the
readability of the gp_toolkit.sql file.

Also update for the split of the column 'waiting' to 'wait_event{,_type} of
pg_stat_activity.

Minor cosmetic changes too.