Fix "could not find pathkey item to sort" error with subqueries.
At least 'appendonly' test was failing because of this. If there is an
aggregate, but no GROUP BY, then there's no sort order to preserve. We
tried to create Motion that preserved the order of the sub-path, but
sub-path's target list might not contain the ordering columns, if they're
not needed in the upper nodes.