Fix merge of upstream commit d8411a6c8b.
We were still throwing the error that the upstream commit relaxed. The
upstream test case added in that comit was failing.

Note: This removes the GPDB-specific change of converting the HeapTuple
returned by the function to a MemTuple. It doesn't seem like a good idea
to do the conversion so eagerly, might as well store the HeapTuples in
the tuplestore, and let them be converted lazily later, if needed.