Do not leak snapshots while creating heap relation indexes.
This commit is left shy of actually going all the way and postpones the actual
solution for the future. The main reason for this is the author's shyness to
address the core issue in the heat of the merge process and several tests
failing. We stop leaking snapshots though.

The deviation from upstream was introduced a long time ago (Dec 2009) with a
laconical message 'Add AO Index'. Currently there seems to be little reason for
this deviation and we could stop passing snapshot and OldestXmin to the
IndexBuild* functions. Not much can be said for the passed estate though.