Make plpgsql pass
Because GPDB has plan-time stable function inlining, there are a few
explain diffs which we resolved.

Also, a few tests added to cover 0426f349effb6bd--a commit related to the
type of message bubbled up in a plpgsql CONTEXT message happened to use
triggers which aren't supported in GPDB and they weren't erroring out in
the same way that others were. We thought about looking into this,
however, it is not easy to tell what the expected error even should be
for these tests. It may or may not be a bug. Thus, we added a FIXME,
just in case.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>