Accept plan changes for tiny changes in cost estimates.
You still get the old plans with "enable_sort=off", but the cost estimates
have slightly changed so that with the defaults, you get the new plans.
The costs are very very close, in fact, the total cost of the hash agg
plan is slightly smaller, but the sort+agg plans have a slightly smaller
startup cost, so the planner prefers them.