Fix estimates on # of rows on aggregates.
We multiplied the number of group by the number of segments, which worked
out for aggregates that run in parallel. But it's not right for aggregates
that run on a single node. This difference in cost estimate changed the
plan in 'aggregates' test.

We're now getting a failure in 'gp_aggregate_costs', where the planner
chooses a sort+agg instead of HashAgg. Not sure why.