Accept plan changes in 'gangsize' tests.
Some of the tests were now choosing GroupAggregates, because it appeared
as cheap as the HashAggregate, when compared using the "fuzzy" factor in
add_path(), because the cost was dominated by the disable cost, because
they could only be implemented using Nested Loop joins. Enable nested
loop joins, to avoid the disable costs.

Some of the aggregates were now distributing to all segments, not only
those where the underlying tables had data. That seems OK and desired;
if you have a 10 segment cluster, but a table only has data on, say,
9 segments, surely you want to utilize the CPU power of the tenth server,
too, when possible.