Adjust tablesample tests to match greenplum
Tablesampling is based in tuple or page sampling. Since in greenplum tuples are
distributed and sampling is performed per segment, we have to cheat a bit to
make certain that the tuples are ending up in the same segment. In practice this
should not matter at all since percentage is cummutative, however for very small
tables as present in the tests, it can create flakyness.

Also includes minor merge cleanup in nodeSamplescan.c