Move GPDB-specific functions to separate file.
The immediate reason to do this now is that the mock test was failing
to compile, because of the Assert that was added to dumputils.c. I'm not
sure why, but separating the GPDB functions to a separate file seem like
a good idea anyway, so let's just do that.