Add merge-specific generated pipeline yml file
The current pipeline that you see here
https://pg-merge.ci.gpdb.pivotal.io/teams/main/pipelines/postgres_merge_REL9_6
is set using this pipeline yml
concourse/pipelines/gpdb_pg_merge.yml

It includes updates for concourse 5.6 as well as some helpful features
for this merge -- long-running jobs have a timeout added and we also
added privileged mode to those tasks so we can run gdb in those
containers (note you still need to install gdb on the container after
hijacking it).

Our intent was to create a merge-specific template in
gpdb-postgres-merge/concourse/pipelines/templates/ which could be used
with the gen_pipeline.py script.
The reason we wanted to put it in the template was that we know
that some people like to set their own pipelines with subsets of the
jobs and such to use while working on the merge. If we only modified the
generated file and checked it in, it would be harder for them to do that
and keep the configuration relatively in sync with the main merge
pipeline. We didn't do this because currently the gen_pipeline.py script
doesn't give you the right `fly` command (different repo and secrets
file override required). We could make a pull request against master to
change this, but, we thought it would be easier for the duration of the
merge to just create a new checked-in generated yml file.

We must remember to decide whether or not we would like to commit this
file to the main repo when we do the merge commit, however. Most likely
we don't want to.
Remember not to try and regenerate this file with gen_pipeline.py

And use this fly command to set the pipeline again with this pipeline
yml file:

 fly -t pg_merge \
     set-pipeline \
     -p postgres_merge_REL9_6 \
     -c ~/workspace/gpdb-postgres-merge/concourse/pipelines/gpdb_pg_merge.yml \
     -l ~/workspace/gp-continuous-integration/secrets/gpdb_common-ci-secrets.yml \
     -l ~/workspace/gp-continuous-integration/secrets/gpdb_master-ci-secrets.dev.yml \
     -l ~/workspace/gp-continuous-integration/secrets/gpdb_merge-ci-secrets.yml \
     -l ~/workspace/gp-continuous-integration/secrets/ccp_ci_secrets_pg_merge.yml \
     -v pipeline-name=postgres_merge_REL9_6

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>